chatbot/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.0
httpx==0.27.0
pillow==10.2.0

# LangChain ecosystem
langchain==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0

# Async support
aiofiles==23.2.1

# Pydantic for FastAPI
pydantic==2.10.2

